Add routes to anecdotes (ex 114)
This might be a really weird solution (passing the different pages to Menu as children and then iterating over them to create a Link and a Route for each separately), but I really wanted to implement it in a way that would require the least refactoring possible when adding a new page. I tried to avoid the situation where one would have to add a link to page manually when creating the component.

Also I hate that you have to iterate over the children twice, but otherwise the contents will be rendered directly after the link, which is not what we want. I'm not sure if accessing childrens' props is best practice either. I basically wanted to see if this was possible.

I already implemented the exercise once by passing the site content to menu as props (a list containing components, thanks javascript), but I found providing the components as children to be far more clear. I think the end product is quite readable, but I'm not sure if conceptually it would be very coherent for someone else that the site content is subject to Menu.